Download search index

Finds and downloads the Elasticsearch index database necessary to set up Photon locally.

Usage

download_searchindex(
  path = ".",
  country = "Monaco",
  date = "latest",
  exact = FALSE,
  only_url = FALSE,
  quiet = FALSE
)

Arguments

path

Path to a directory where the identified file should be stored.

country

Character string that can be identified by countryname as a country. An extract for this country will be downloaded. If "planet", downloads a global search index.

date

Character string or date-time object used to specify the creation date of the search index. If "latest", will download the file tagged with "latest". If a character string, the value should be parseable by as.POSIXct. If exact = FALSE, the input value is compared to all available dates and the closest date will be selected. Otherwise, a file will be selected that exactly matches the input to date.

exact

If TRUE, exactly matches the date. Otherwise, selects the date with lowest difference to the date parameter.

only_url

If TRUE, downloads the search index. Otherwise, only returns a link to the file.

quiet

If TRUE, suppresses all informative messages.

Value

If only_url = FALSE, returns the local path to the downloaded file. Otherwise, returns the URL to the remote file.

Note

Depending on the country, search index databases tend to be very large. The global search index is about 75 GB of size (10/2024). Keep that in mind when running this function.

Examples

# download the latest extract of Monaco
download_searchindex(path = tempdir())
#> ℹ Fetching search index for Monaco, created on latest
#> ✔ Successfully downloaded search index. [1.6s]
#> 
#> [1] "/tmp/RtmpYxh4qb/photon-db-mc-latest.tar.bz2"

# download the latest extract of American Samoa
download_searchindex(path = tempdir(), country = "Samoa")
#> ℹ Fetching search index for Samoa, created on latest
#> ✔ Successfully downloaded search index. [158ms]
#> 
#> [1] "/tmp/RtmpYxh4qb/photon-db-ws-latest.tar.bz2"

# download an extract from a month ago
download_searchindex(path = tempdir(), country = "Monaco", date = Sys.time() - 2629800)
#> ℹ Fetching search index for Monaco, created on 2024-10-20
#> ✔ Successfully downloaded search index. [158ms]
#> 
#> [1] "/tmp/RtmpYxh4qb/photon-db-mc-241020.tar.bz2"

# if possible, download an extract from today
try(download_searchindex(path = tempdir(), country = "Monaco", date = Sys.Date(), exact = TRUE))
#> Error in download_searchindex(path = tempdir(), country = "Monaco", date = Sys.Date(),  : 
#>   ! Specified `date` does not match any available dates.
#> ℹ Consider setting `exact = FALSE`.

# get the latest global coverage
# NOTE: the file to be downloaded is several tens of gigabytes of size!
if (FALSE) {
  download_searchindex(path = tempdir(), country = "planet")
}