photon: Interface to the 'Photon' geocoding API

Features unstructured, structured and reverse geocoding using the 'Photon' geocoding API. Facilitates the setup of local Photon instances to enable offline geocoding.

Terms of use

From https://photon.komoot.io on using the public API:

"You can use the API for your project, but please be fair - extensive usage will be throttled. We do not guarantee for the availability and usage might be subject of change in the future."

Note that these terms only apply to the public API (new_photon()), and not to local instances (e.g. new_photon(path = "."))! For the public API, the package sets a default of 1 request per second (see below).

Global options

A number of global options can be set that change the behavior of package functions. These include:

photon_throttle

Rate limit used to throttle requests. By default, no throttle is set for non-komoot instances. For komoot's public API, this option defaults to 1 request per second. See req_throttle.

photon_max_tries

Number of retries a failing request should do before ultimately aborting. Defaults to 3. See req_retry.

photon_debug

Whether to echo the command of external processes and GET requests sent to photon. Defaults to FALSE.

Author

Maintainer: Jonas Lieth jonas.lieth@gesis.org [copyright holder]