Local photon instance

This R6 class is used to initialize and manage local photon instances. It can download and setup the Java, the photon executable, and the necessary Elasticsearch search index. It can start, stop, and query the status of the photon instance. It is also the basis for geocoding requests at it is used to retrieve the URL for geocoding.

Super class

photon::photon -> photon_local

Public fields

path

Path to the directory where the photon instance is stored.

proc

process object that handles the external process running photon.

Methods

Public methods

photon_local$new()

photon_local$info()

photon_local$purge()

photon_local$start()

photon_local$stop()

photon_local$is_running()

photon_local$is_ready()

photon_local$get_url()

photon_local$clone()

Method new()

Initialize a local photon instance. If necessary, downloads the photon executable, the search index, and Java.

Usage

photon_local$new(
  path = "./photon",
  photon_version = NULL,
  country = NULL,
  date = "latest",
  exact = FALSE,
  quiet = FALSE
)

Arguments

path

Path to a directory where the photon executable and data should be stored. Defaults to a directory "photon" in the current working directory.

photon_version

Version of photon to be used. A list of all releases can be found here: https://github.com/komoot/photon/releases/.

country

Character string that can be identified by countryname as a country. An extract for this country will be downloaded. If NULL, downloads a global search index.

date

Character string or date-time object used to specify the creation date of the search index. If "latest", will download the file tagged with "latest". If a character string, the value should be parseable by as.POSIXct. If exact = FALSE, the input value is compared to all available dates and the closest date will be selected. Otherwise, a file will be selected that exactly matches the input to date.

exact

If TRUE, exactly matches the date. Otherwise, selects the date with lowest difference to the date parameter.

quiet

If TRUE, suppresses all informative messages.

Method info()

Retrieve metadata about the java and photon version used as well as the country and creation date of the Eleasticsearch search index.

Usage

photon_local$info()

Method purge()

Kill the photon process and remove the directory. Useful to get rid of an instance entirely.

Usage

photon_local$purge()

Method start()

Start a local instance of the Photon geocoder. Runs the jar executable located in the instance directory.

Usage

photon_local$start(
  min_ram = 5,
  max_ram = 10,
  host = "0.0.0.0",
  port = "2322",
  ssl = FALSE,
  java_options = NULL,
  photon_options = NULL
)

Arguments

min_ram

Initial RAM to be allocated to the Java process (-Xms flag).

max_ram

Maximum RAM to be allocated to the Java process (-Xmx flag)

host

Character string of the host name that the geocoder should be opened on.

port

Port that the geocoder should listen to.

ssl

If TRUE, uses https, otherwise http. Defaults to FALSE.

java_options

List of further flags passed on to the java command.

photon_options

List of further flags passed on to the photon jar in the java command.

Method stop()

Kills the running photon process.

Usage

photon_local$stop()

Method is_running()

Checks whether the photon instance is running and ready. The difference to $is_ready() is that $is_running() checks specifically if the running photon instance is managed by a process from its own photon object. In other words, $is_running() returns TRUE if both $proc$is_alive() and $is_ready() return TRUE. This method is useful if you want to ensure that the photon object can control its photon server (mostly internal use).

Usage

photon_local$is_running()

Method is_ready()

Checks whether the photon instance is ready to take requests. This is the case if the photon server returns a HTTP 400 when sending a queryless request. This method is useful if you want to check whether you can send requests.

Usage

photon_local$is_ready()

Method get_url()

Constructs the URL that geocoding requests should be sent to.

Usage

photon_local$get_url()

Method clone()

The objects of this class are cloneable with this method.

Usage

photon_local$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# start a new instance using a Monaco extract
photon <- new_photon(path = tempdir(), country = "Monaco")

# start a new instance with an older photon version
photon <- new_photon(path = tempdir(), photon_version = "0.4.1")

# start a new instance with a specific java version
photon <- new_photon(path = tempdir(), java_version = 17)
}