Local photon instance

This R6 class is used to initialize and manage local photon instances. It can download and setup the Java, the photon executable, and the necessary Elasticsearch search index. It can start, stop, and query the status of the photon instance. It is also the basis for geocoding requests at it is used to retrieve the URL for geocoding.

ElasticSearch

The standard version of photon uses ElasticSearch indices to geocode. These search indices can be self-provided by importing an existing Nominatim database or they can be downloaded from the Photon download server. Use nominatim = TRUE to indicate that no ElasticSearch indices should be downloaded. See vignette("nominatim-import", package = "photon") for details on how to import from Nominatim.

OpenSearch

To enable structured geocoding, the photon geocoder needs to be built to support OpenSearch. Currently, this feature is only experimental and needs to be manually built using gradle. When using the OpenSearch version of photon, no pre-built ElasticSearch indices can be used and the data must be directly imported from a Nominatim database. Refer to the photon README for details. These steps cannot be done by {photon} and must be done before initializing photon_local. If an OpenSearch jar is detected, the nominatim parameter is forced to TRUE.

Super class

photon::photon -> photon_local

Public fields

path

Path to the directory where the photon instance is stored.

proc

process object that handles the external process running photon.

Methods

Public methods

photon_local$new()

photon_local$mount()

photon_local$info()

photon_local$purge()

photon_local$import()

photon_local$start()

photon_local$stop()

photon_local$is_running()

photon_local$is_ready()

photon_local$get_url()

photon_local$clone()

Method new()

Initialize a local photon instance. If necessary, downloads the photon executable, the search index, and Java.

Usage

photon_local$new(
  path = "./photon",
  photon_version = NULL,
  country = NULL,
  date = "latest",
  exact = FALSE,
  opensearch = FALSE,
  overwrite = FALSE,
  quiet = FALSE
)

Arguments

path

Path to a directory where the photon executable and data should be stored. Defaults to a directory "photon" in the current working directory.

photon_version

Version of photon to be used. A list of all releases can be found here: https://github.com/komoot/photon/releases/. Ignored if jar is given. If NULL, uses the latest known version.

country

Character string that can be identified by countryname as a country. An extract for this country will be downloaded. If NULL, downloads a global search index.

date

Character string or date-time object used to specify the creation date of the search index. If "latest", will download the file tagged with "latest". If a character string, the value should be parseable by as.POSIXct. If exact = FALSE, the input value is compared to all available dates and the closest date will be selected. Otherwise, a file will be selected that exactly matches the input to date.

exact

If TRUE, exactly matches the date. Otherwise, selects the date with lowest difference to the date parameter.

opensearch

If TRUE, looks for an OpenSearch version of photon in the specified path. Opensearch-based photon supports structured geocoding queries but has to be built manually using gradle. Hence, it cannot be downloaded directly. If no OpenSearch executable is found in the search path, then this parameter is set to FALSE. Defaults to FALSE.

overwrite

If TRUE, overwrites existing jar files and search indices when initializing a new instance. Defaults to FALSE.

quiet

If TRUE, suppresses all informative messages.

Method mount()

Attach the object to the session. If mounted, all geocoding functions send their requests to the URL of this instance. Manually mounting is useful if you want to switch between multiple photon instances.

Usage

photon_local$mount()

Method info()

Retrieve metadata about the java and photon version used as well as the country and creation date of the Eleasticsearch search index.

Usage

photon_local$info()

Method purge()

Kill the photon process and remove the directory. Useful to get rid of an instance entirely.

Usage

photon_local$purge(ask = TRUE)

Arguments

ask

If TRUE, asks for confirmation before purging the instance.

Method import()

Import a Postgres Nominatim database to photon. Runs the photon jar file using the additional parameter -nominatim-import. Requires a running Nominatim database that can be connected to.

Usage

photon_local$import(
  host = "127.0.0.1",
  port = 5432,
  database = "nominatim",
  user = "nominatim",
  password = "",
  structured = FALSE,
  update = FALSE,
  enable_update_api = FALSE,
  languages = c("en", "fr", "de", "it"),
  countries = NULL,
  extra_tags = NULL,
  json = FALSE,
  timeout = 60,
  java_opts = NULL,
  photon_opts = NULL
)

Arguments

host

Postgres host of the database. Defaults to "127.0.0.1".

port

Postgres port of the database. Defaults to 5432.

database

Postgres database name. Defaults to "nominatim".

user

Postgres database user. Defaults to "nominatim".

password

Postgres database password. Defaults to "".

structured

If TRUE, enables structured query support when importing the database. This allows the usage of structured. Structured queries are only supported in the OpenSearch version of photon. See section "OpenSearch" above. Defaults to FALSE.

update

If TRUE, fetches updates from the Nominatim database, updating the search index without offering an API. If FALSE, imports the database an deletes the previous index. Defaults to FALSE.

enable_update_api

If TRUE, enables an additional endpoint /nominatim-update, which allows updates from Nominatim databases.

languages

Character vector specifying the languages to import from the Nominatim databases. Defaults to English, French, German, and Italian.

countries

Character vector specifying the country codes to import from the Nominatim database. Defaults to all country codes.

extra_tags

Character vector specifying extra OSM tags to import from the Nominatim database. These tags are used to augment geocoding results. Defaults to NULL.

json

If TRUE, dumps the imported Nominatim database to a JSON file and returns the path to the output file. Defaults to FALSE.

timeout

Time in seconds before the java process aborts. Defaults to 60 seconds.

java_opts

List of further flags passed on to the java command.

photon_opts

List of further flags passed on to the photon jar in the java command. See cmd_options for a helper function to import external Nominatim databases.

Method start()

Start a local instance of the Photon geocoder. Runs the jar executable located in the instance directory.

Usage

photon_local$start(
  host = "0.0.0.0",
  port = "2322",
  ssl = FALSE,
  timeout = 60,
  java_opts = NULL,
  photon_opts = NULL
)

Arguments

host

Character string of the host name that the geocoder should be opened on.

port

Port that the geocoder should listen to.

ssl

If TRUE, uses https, otherwise http. Defaults to FALSE.

timeout

Time in seconds before the java process aborts. Defaults to 60 seconds.

java_opts

List of further flags passed on to the java command.

photon_opts

List of further flags passed on to the photon jar in the java command. See cmd_options for a helper function to import external Nominatim databases.

Details

While there is a certain way to determine if a photon instance is ready, there is no clear way as of yet to determine if a photon setup has failed. Due to this, a failing setup is mostly indicated by the setup hanging after emitting a warning. In this case, the setup has to be interrupted manually.

Method stop()

Kills the running photon process.

Usage

photon_local$stop()

Method is_running()

Checks whether the photon instance is running and ready. The difference to $is_ready() is that $is_running() checks specifically if the running photon instance is managed by a process from its own photon object. In other words, $is_running() returns TRUE if both $proc$is_alive() and $is_ready() return TRUE. This method is useful if you want to ensure that the photon object can control its photon server (mostly internal use).

Usage

photon_local$is_running()

Method is_ready()

Checks whether the photon instance is ready to take requests. This is the case if the photon server returns a HTTP 400 when sending a queryless request. This method is useful if you want to check whether you can send requests.

Usage

photon_local$is_ready()

Method get_url()

Constructs the URL that geocoding requests should be sent to.

Usage

photon_local$get_url()

Method clone()

The objects of this class are cloneable with this method.

Usage

photon_local$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
dir <- file.path(tempdir(), "photon")

# start a new instance using a Monaco extract
photon <- new_photon(path = dir, country = "Monaco")

# start a new instance with an older photon version
photon <- new_photon(path = dir, photon_version = "0.4.1")

# import a nominatim database using OpenSearch photon
# this example requires the OpenSearch version of photon and a running
# Nominatim server.
file.copy(
  "path/to/photon-opensearch-0.5.0.jar",
  file.path(dir, "photon-opensearch-0.5.0.jar")
)
photon <- new_photon(path = dir, nominatim = TRUE)
photon$start(photon_options = cmd_options(port = 29146, password = "pgpass"))
}