Structured geocoding

Geocode a set of place information such as street, house number, or post code. Structured geocoding is generally more accurate but requires more information than unstructured geocoding. Note that structured geocoding must be specifically enabled when building a Nominatim database. It is generally not available on komoot's public API and on pre-built search indices through download_searchindex.

Usage

structured(
  .data,
  limit = 3,
  lang = "en",
  bbox = NULL,
  osm_tag = NULL,
  layer = NULL,
  locbias = NULL,
  locbias_scale = NULL,
  zoom = NULL,
  progress = interactive()
)

Arguments

.data

Dataframe or list containing structured information on a place to geocode. Can contain the columns street, housenumber, postcode, city, district, county, state, and countrycode. At least one of these columns must be present in the dataframe.

limit

Number of results to return. Defaults to 3.

lang

Language of the results.

bbox

Any object that can be parsed by st_bbox. Results must lie within this bbox.

osm_tag

Character string giving an OSM tag to filter the results by. See details.

layer

Character string giving a layer to filter the results by. Can be one of "house", "street", "locality", "district", "city", "county", "state", "country", or "other".

locbias

Numeric vector of length 2 or an sfg point geometry that specifies a location bias for geocoding. Geocoding results are biased towards this point. The radius of the bias is controlled through zoom and its prominence through location_bias_scale.

locbias_scale

Numeric vector specifying the prominence of point in location_bias. Defaults to 0.2.

zoom

Numeric vector specifying the radius of the location_bias. Corresponds to the zoom level in OpenStreetMap. Defaults to 16.

progress

If TRUE, shows a progress bar for longer queries.

Details

Filtering by OpenStreetMap tags follows a distinct syntax explained on https://github.com/komoot/photon. In particular:

Include places with tag: key:value

Exclude places with tag: !key:value

Include places with tag key: key

Include places wih tag value: :value

Exclude places with tag key: !key

Exclude places with tag value: :!value